requirements.txt
anyio==4.3.0
appdirs==1.4.4
asyncio==3.4.3
attrs==23.2.0
bottle==0.12.25
cached-property==1.5.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.7
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
lxml==5.2.2
platformdirs==4.2.2
protobuf==5.26.1
psycopg2-binary==2.9.9
pycparser==2.22
python-environ==0.4.54
python-telegram-bot==21.1.1
pytz==2024.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
tornado==6.4
ttictoc==0.5.6
urllib3==2.2.1
zeep==4.2.1
